+++ renderapi/transform/ThinPlateSpline.py
@@ -1,7 +1,7 @@
 import numpy as np
 from ..errors import RenderError
 from ..utils import encodeBase64, decodeBase64
-from .generalTransforms import Transform
+from .transform import Transform
 
 
 class ThinPlateSplineTransform(Transform):
@@ -120,7 +120,9 @@ class ThinPlateSplineTransform(Transform):
         else:
             b64_1 = "null"
 
-        blk2 = np.concatenate((self.srcPts.flatten(order='F'), self.dMtxDat.flatten(order='C')))
+        blk2 = np.concatenate((
+            self.srcPts.flatten(order='F'),
+            self.dMtxDat.flatten(order='C')))
         b64_2 = encodeBase64(blk2)
 
         return '{} {} {}'.format(header, b64_1, b64_2)
